Removed warnings

+++ TricksterEngine/src/Window.h
@@ -47,7 +47,7 @@ namespace Trickster {
 		virtual void GetCursorPos(double* x, double* y) = 0;
 		virtual bool GetClick(int MouseKey) = 0;
 		virtual void SetClick(int MouseKey, bool value) = 0;
-		virtual void Resize(float a_Width, float a_Height) = 0;
+		virtual void Resize(int a_Width, int a_Height) = 0;
 		
 		static Window* Create(const WindowProps& props = WindowProps());
 	protected:
